commit before messing with css

+++ src/components/DeleteCommentButton.jsx
@@ -1,22 +1,26 @@
 import React from "react";
 import axios from "axios";
 
-export default function DeleteCommentButton({ comment_id, deleteComment }) {
-  console.log(comment_id, "<<<<<<<<");
+export default function DeleteCommentButton({
+  comment_id,
+  deleteComment,
+  currentUser,
+  author
+}) {
 
   const handleDeleteClick = event => {
-    console.log("handle delete click function", comment_id);
-
-    return axios
-      .delete(
-        `https://amelias-news-api.herokuapp.com/api/comments/${comment_id}`
-      )
-      .then(() => deleteComment(comment_id))
-      .catch(err => {
-        this.setState({
-          err: { status: err.response.status, msg: err.response.data.msg }
+    if (currentUser === author ) {
+      return axios
+        .delete(
+          `https://amelias-news-api.herokuapp.com/api/comments/${comment_id}`
+        )
+        .then(() => deleteComment(comment_id))
+        .catch(err => {
+          this.setState({
+            err: { status: err.response.status, msg: err.response.data.msg }
+          });
         });
-      });
+    }
   };
 
   return (
